Fix Electron Module References
When running electron-cucumber, support code would fail because the
electron module could not be found. The NODE_PATH environment variable
that is passed to the Electron child process needs to be set to the
path of Electron’s API within the electron.asar package. I fixed this.

I created a sample set of feature tests to verify that Cucumber runs
within electron. The sample creates a BrowserWindow and then uses
webContents to execute JavaScript to populate the fields of a simple
sign-in form.